refactor: modularize analytics and csv parsing
